Softened some of the wording used
Updated the intro line, and removed mention of "calling out".
A user on Reddit described this as a hit-list, which is really not my
intent nor how I want this to be percieved.
Updating the intro to sound less accusatory, and be more factual about
what is being documented here, while removing "calling out" wording
since that could be taken as a desire for active action of some kind
which again is not my intent, since I don't expect community/user
actioning of any kind, but more in regard to documenting these scenarios
so they are not invisible.